drop python-xlib
Python-xlib misses some essential features like shm support.
This is why the Xshm module was written,
but this leads to redundant dependencies.
So this commit removes python-xlib as dependency.

During the translation of the code the determination of the pid of a
window was also improved.
The XRes extension is now used to determine the pid of a window.